requirements.txt
ruamel.yaml
rasterio
rio-cogeo
numpy~=1.15
boto3
uuid
